ExceptionFactories namespace => Exceptions namespace
