+++ DbC.Net.TestAndExampleResources/GlobalUsings.cs
@@ -1,4 +1,4 @@
 ﻿global using System.Diagnostics.CodeAnalysis;
 
-global using DbC.Net.ExceptionFactories;
+global using DbC.Net.Exceptions;
 global using DbC.Net.Transforms;
