+++ DbC.Net.Tests.Benchmarks/GlobalUsings.cs
@@ -1,7 +1,7 @@
 ﻿global using BenchmarkDotNet.Attributes;
 global using BenchmarkDotNet.Running;
 
-global using DbC.Net.ExceptionFactories;
+global using DbC.Net.Exceptions;
 global using DbC.Net.FloatingPoint;
 global using DbC.Net.TestAndExampleResources;
 global using DbC.Net.Tests.Benchmarks;
