+++ DbC.Net.Tests.Unit/DbC.Net.Tests.Unit.csproj
@@ -27,4 +27,8 @@
     <ProjectReference Include="..\DbC.Net\DbC.Net.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="ExceptionFactories\" />
+  </ItemGroup>
+
 </Project>
