+++ DbC.Net.Tests.Unit/GlobalUsings.cs
@@ -2,7 +2,6 @@ global using System.Collections;
 global using System.Numerics;
 
 global using DbC.Net.Exceptions;
-global using DbC.Net.ExceptionFactories;
 global using DbC.Net.FloatingPoint;
 global using DbC.Net.TestAndExampleResources;
 global using DbC.Net.Tests.Unit.TestData;
