+++ DbC.Net/DbC.Net.csproj
@@ -52,4 +52,8 @@
 	  </None>
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="ExceptionFactories\" />
+	</ItemGroup>
+
 </Project>
