+++ DbC.Net/GlobalUsings.cs
@@ -2,6 +2,5 @@
 global using System.Runtime.CompilerServices;
 
 global using DbC.Net.Exceptions;
-global using DbC.Net.ExceptionFactories;
 global using DbC.Net.FloatingPoint;
 global using DbC.Net.Transforms;
